FIX:Comparison warning
Resovle warning comparision between strlen() and int.
Correct (int)(strlen()).

+++ Code.c
@@ -304,7 +304,7 @@ void showLeaderBoard(void)
 	while (!feof(leaderboard)) 
 	{
 		fscanf(leaderboard, "%s %s %s", player1_name, player2_name, result);
-		if (strlen(player1_name) < tab_size)
+		if ((int)(strlen(player1_name) )< tab_size)
 		{
 			printf("|\t%s\t\t|", player1_name);
 		}
@@ -313,7 +313,7 @@ void showLeaderBoard(void)
 			printf("|\t%s\t|", player1_name);
 		}
 
-		if (strlen(player2_name) < tab_size) 
+		if ((int)(strlen(player2_name)) < tab_size) 
 		{
 			printf("\t%s\t\t|", player2_name);
 		}
@@ -322,7 +322,7 @@ void showLeaderBoard(void)
 			printf("\t%s\t|", player2_name);
 		}
 
-		if (strlen(result) < tab_size) 
+		if ((int)(strlen(result) )< tab_size) 
 		{
 			printf("\t%s\t\t|\n", result);
 		}
